Use tdewolff/minify for production builds
Minifying content is hard and involves a lot of edge cases which I don't
want to handle.

To get 99% of the way there, I now use a third party library to perform
most of the minification, and then use a custom method to undergo the
last 1% minification that are specific optimizations to 2web.

I was able to shave off an additional 67 bytes (~8.9%) off the benchmark
build!