Add vanilla HTML + JS benchmarks
I will never be able to get exactly this score, but the closer I can get
the better.

Note: That the vanilla benchmark is optimized for size, and not reactive
speed. It actually has a less than optimal first load time because it's
reactivity is hoisted on the first render.