Re-use main thread for last parallel iteration
Spinning up a new thread takes time.
However, because a parallel for loop is blocking, we can re-use the main
thread for the last operation meaning that we don't have to create a new
thread.

Additionally, because we compile imports in parallel, any scripts that
import one component now become synchronous.